+++ src/routes/index.ts
@@ -1,12 +1,9 @@
 import users from "./users";
-import thread from "./thread";
+import thread from "./threads";
 import { FastifyInstance, FastifyPluginOptions } from "fastify";
 import me from "./me";
 import categories from "./categories";
-import category from "./category";
 import notifications from "./notifications/notifications";
-import threads from "./threads";
-import user from "./user";
 
 export default (
     fastify: FastifyInstance,
@@ -14,12 +11,9 @@ export default (
     done: (e?: Error) => void
 ) => {
     fastify.register(users, { prefix: "/users" });
-    fastify.register(user, { prefix: "/user" });
     fastify.register(me, { prefix: "/me" });
     fastify.register(categories, { prefix: "/categories" });
-    fastify.register(category, { prefix: "/category" });
-    fastify.register(thread, { prefix: "/thread" });
-    fastify.register(threads, { prefix: "/threads" });
+    fastify.register(thread, { prefix: "/threads" });
     fastify.register(notifications, { prefix: "/notifications" });
     done();
 };